+++ docs/changelog.rst
@@ -1,6 +1,7 @@
 Next version (not yet released)
 -------------------------------
 
+- Add Oracle Instant Client to runtime dependencies.
 
 
 Grocker 2.3.1 (2016-03-03)