+++ docs/changelog.rst
@@ -3,6 +3,7 @@ Next version (not yet released)
 
 - Only install needed runtime in images.
 - Allow to set system dependencies by project.
+- Remove dependencies to host UID
 
 
 Grocker 2.3.1 (2016-03-03)